% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthGOF.R
\name{earth_GOF}
\alias{earth_GOF}
\title{Goodness of fitting values of an earth model.}
\usage{
earth_GOF(model)
}
\arguments{
\item{model}{an earth model (keepxy argument should be set TRUE).}
}
\value{
The goodness of fitting  values of \code{model}.
}
\description{
Goodness of fitting values of an earth model.
}
\examples{
# an example for normal lm earth model

library(earth)
library(earthInfo)
data("iris")
fit <- earth(Sepal.Width~., data=iris, degree=2,keepxy = TRUE)
gof <- earth_GOF(fit)
summary(gof,digits = 3)



# an example for a binomial glm earth model

library(earth)
library(earthInfo)
data(etitanic)
fit <- earth(survived~., data=etitanic, degree =2, glm = list(family=binomial),keepxy = TRUE)
gof <- earth_GOF(fit)
summary(gof)
}
