% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGCV.R
\name{fit_GCV}
\alias{fit_GCV}
\title{Small function to minimize GCV value of an earth model.}
\usage{
fit_GCV(formula, data, max_degree = 3, ...)
}
\arguments{
\item{formula}{standart R formula (Eg: Spescies~ .).}

\item{data}{data frame .}

\item{max_degree}{using to tune model.}
}
\value{
Returns a list consists minimum GCV value, maximum interactions and optimum degree of the model.
}
\description{
Small function to minimize GCV value of an earth model.
}
\examples{
library(earthInfo)
data("iris")
fit <- fitGCV(Sepal.Width~., data=iris, maxDegree = 10)
}
