% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Small function to maximize  ....  of an binomial glm earth model.}
\usage{
fit_glm(formula, data, max_degree = 3, max_prune = 30,
  method = "acc", ...)
}
\arguments{
\item{formula}{standart R formula (Eg: Spescies~ .).}

\item{data}{data frame .}

\item{max_degree}{using to tune model.}

\item{max_prune}{using to tune model.}

\item{method}{use "acc" for accuracy  or "auc" for are under the curve.}
}
\value{
Returns a list consists maximum ac value, maximum interactions and optimum degree of the model.
}
\description{
Small function to maximize  ....  of an binomial glm earth model.
}
\examples{
library(earthInfo)
data("iris")
fit <- fit_GCV(Sepal.Width~., data=iris, maxDegree = 10)
}
