% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fitGlm}
\alias{fitGlm}
\title{Small function to maximize  accuracy of an binomial glm earth model.}
\usage{
fitGlm(formula, data, maxDegree = 9, method = "acc", ...)
}
\arguments{
\item{formula}{standart R formula (Eg: Spescies~ .).}

\item{data}{data frame .}

\item{maxDegree}{using to tune model.}

\item{method}{use "acc" for accuracy  or "auc" for are under the curve.}
}
\value{
Returns a list consists maximum ac value, maximum interactions and optimum degree of the model.
}
\description{
Small function to maximize  accuracy of an binomial glm earth model.
}
\examples{
library(earthInfo)
data(etitanic)
fit <- fitGlm(survived~., data=etitanic, maxDegree = 10)
}
